requirements.txt
ultralytics
torch
opencv-python
numpy
Pillow
diffusers
transformers
streamlit
